修复，修改，增强，完善--妙笔生花 修复了虫群挑战下BOSS掉两把钥匙的问题 修复了魔法纸能够被升级的问题 修复了进入萨尔贡后其他新建存档中出现巨型蘑菇的问题 修复了宝箱怪被奥能射击的真伤击杀时不掉落物品的问题 我记得好像修复了诅咒固化剂导致其他诅咒法杖也固化的问题 其他的忘记了 艺术打磨现在会为玩家恢复生命了 增强了寒气词条 初步完成了小绿的妙笔生花

+++ Arrrknights/core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Goo.java
@@ -272,7 +272,7 @@ public class Goo extends Mob {
 		Dungeon.level.unseal();
 		
 		GameScene.bossSlain();
-		Dungeon.level.drop( new SkeletonKey( Dungeon.depth ), pos ).sprite.drop();
+		if(!this.isSwarmChild)Dungeon.level.drop( new SkeletonKey( Dungeon.depth ), pos ).sprite.drop();
 		Dungeon.level.drop(new Certificate(2), pos).sprite.drop(pos);
 		
 		//60% chance of 2 blobs, 30% chance of 3, 10% chance for 4. Average of 2.5
