+++ Arrrknights/core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/King.java
@@ -150,7 +150,7 @@ public class King extends Mob {
 
 		GameScene.bossSlain();
 		Dungeon.level.drop( new ArmorKit(), pos ).sprite.drop();
-		Dungeon.level.drop( new SkeletonKey( Dungeon.depth ), pos ).sprite.drop();
+		if(!this.isSwarmChild)Dungeon.level.drop( new SkeletonKey( Dungeon.depth ), pos ).sprite.drop();
 		
 		super.die( cause );
